Extract the BadgeLabel to a separate file. Remove the ducplicated custom view
