+++ WooCommerce/Classes/ViewRelated/OrdersBadgeController.swift
@@ -60,115 +60,3 @@ private extension OrdersBadgeController {
         static let height      = CGFloat(17)
     }
 }
-
-
-
-// MARK: - Badge UIView
-//
-private class BadgeView: UIView {
-
-    private var borderWidth = CGFloat(1) // Border line width defaults to 1
-
-    /// Designated Initializer
-    ///
-    init(frame: CGRect, borderWidth: CGFloat) {
-        super.init(frame: frame)
-        self.borderWidth = borderWidth
-        setupSubviews()
-    }
-
-    /// Required Initializer
-    ///
-    required init?(coder aDecoder: NSCoder) {
-        super.init(coder: aDecoder)
-        setupSubviews()
-    }
-
-    private func setupSubviews() {
-        backgroundColor = .clear
-    }
-
-    override func draw(_ rect: CGRect) {
-        let path = UIBezierPath(ovalIn: CGRect(x: rect.origin.x + borderWidth,
-                                               y: rect.origin.y + borderWidth,
-                                               width: rect.size.width - borderWidth*2,
-                                               height: rect.size.height - borderWidth*2))
-        StyleManager.wooAccent.setFill()
-        path.fill()
-
-        path.lineWidth = borderWidth
-        StyleManager.wooWhite.setStroke()
-        path.stroke()
-    }
-}
-
-
-class BadgeLabel: UILabel {
-    @IBInspectable var horizontalPadding: CGFloat = 0 {
-        didSet {
-            invalidateIntrinsicContentSize()
-            setNeedsLayout()
-        }
-    }
-
-    // MARK: Initialization
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        setupView()
-    }
-
-    required init?(coder aDecoder: NSCoder) {
-        super.init(coder: aDecoder)
-        setupView()
-    }
-
-    fileprivate func setupView() {
-        textAlignment = .center
-        layer.masksToBounds = true
-
-        cornerRadius = 2.0
-    }
-
-    // MARK: Padding
-
-    override func drawText(in rect: CGRect) {
-        let insets = UIEdgeInsets.init(top: 0, left: horizontalPadding, bottom: 0, right: horizontalPadding)
-        super.drawText(in: rect.inset(by: insets))
-    }
-
-    override var intrinsicContentSize: CGSize {
-        var paddedSize = super.intrinsicContentSize
-        paddedSize.width += 2 * horizontalPadding
-        return paddedSize
-    }
-
-    // MARK: Computed Properties
-
-    @IBInspectable var borderColor: UIColor {
-        get {
-            return UIColor(cgColor: layer.borderColor!)
-        }
-        set {
-            layer.borderColor = newValue.cgColor
-        }
-    }
-
-    @IBInspectable var borderWidth: CGFloat {
-        get {
-            return layer.borderWidth
-        }
-        set {
-            layer.borderWidth = newValue
-        }
-    }
-
-    @IBInspectable var cornerRadius: CGFloat {
-        get {
-            return layer.cornerRadius
-        }
-        set {
-            layer.cornerRadius = newValue
-        }
-    }
-}
